M_15_Documentation/homework/pyproject.toml
[tool.poetry]
name = "M_15_Documentation"
version = "0.1.0"
description = ""
authors = ["ramazan_aisaev <aisaev-68@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
pillow = "^9.4.0"
python-decouple = "^3.7"
requests = "^2.28.2"
pymemcache = "^4.0.0"
djangorestframework = "^3.14.0"
faker = "^17.6.0"
django-bootstrap4 = "^22.3"
drf-yasg = "^1.21.5"
docutils = "^0.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
